Working on getting 64-bit working.
git-svn-id: svn://svn.code.sf.net/p/opticks-extras/code/branches/future/EXTRAS-1/IDL@21 2d91f92e-8166-0410-902e-d52292f61e86

+++ Code/IdlStart/MetadataCommands.cpp
@@ -16,7 +16,6 @@
 #include <string>
 #include <stdio.h>
 #include <idl_export.h>
-#include <idl_callproxy.h>
 
 /*!@cond INTERNAL */
 namespace
@@ -30,7 +29,7 @@ namespace
    }
 
    template<typename T>
-   DataVariant idl_to_DataVariant(int total, char* pValue)
+   DataVariant idl_to_DataVariant(size_t total, char* pValue)
    {
       T* pValReal = reinterpret_cast<T*>(pValue);
       if (total == 1)
@@ -49,7 +48,7 @@ namespace
    }
 
    template<>
-   DataVariant idl_to_DataVariant<bool>(int total, char* pValue)
+   DataVariant idl_to_DataVariant<bool>(size_t total, char* pValue)
    {
       // special case since bool isn't always 1 byte
       if (total == 1)
